登录功能完善
解决了登录功能卡页的BUG，并同步完成了令牌校验的功能。

+++ back-end-code/ssmp-all/ssmp/ssmp-common/src/main/java/com/eternalrights/properties/JwtProperties.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 @Data
 public class JwtProperties {
     private String userSecretKey;
-    private long userTtl;
+    private Integer userTtl;
     private String userTokenName;
 
 }
