+++ back-end-code/ssmp-all/ssmp/ssmp-pojo/src/main/java/com/eternalrights/vo/AuthResult.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class AuthResult {
     private String token;
-    private long expiresIn;
+    private Integer expiresIn;
     private User user;
     private Boolean success;
 }
